Improvements to tinybootloader to make it suitable for use in my projects - specifically:
It is changed to reside entirely inside the bootblock, and to run when
the PIC is reset.

This means:

- the bootblock can be write protected, making it pretty much
impossible to make a device unusable in the field

- the bootloader is always entered on power up (rather than relying on
the device firmware having a way to enter the bootloader) - this means
you can get into the bootloader (and hence rescue a device in the
field) even if the user software has been completely corrupted

As is commonly done, we remap the interrupt vectors into the space
just after bootloader block so that the user software can continue
to handle interrupts.

Change the delay it waits in the bootloader before going to the user
program to be 0.2 seconds - this is long enough for a PC to get in and
connect to the bootloader, and short enough that it is pretty much
unnoticeable to the end user of a device.

Changes to used CONHFIG, as __CONFIG is now deprecated by MPLAB

Protect all blocks against tablereads, as otherwise a "hacker" could
load on a very small custom firmware with the sole purpose of reading
the rest of the firmware and writing it to the serial port.

Updated include file locations, as we keep everything in the same
directory.

Change baud rate to 57600 (this is arbitary and happens to work better
for my projects).